+++ test/fixtures/context/use_import_etc.ex
@@ -36,3 +36,19 @@ defmodule Traverse.Foo do
     micky(:entenhausen)
   end
 end
+
+defmodule Traverse.Timer do
+  import :timer
+end
+
+defmodule Traverse.RequireAlias do
+  require alias Traverse.Pluto
+end
+
+defmodule Traverse.RequireAliasAs do
+  require alias Traverse.Pluto, as: Foo
+end
+
+defmodule Traverse.AliasRequire do
+  alias require Traverse.Pluto
+end